Use shell=True in subprocess calls on Windows
This sets the value for the `shell=` parameter in subprocess calls to
click's belief whether or not the system is Windows. Setting it to True
alltogether completely breaks subprocess calls to paths with whitespace
on MacOS. This commit also moves the `if force` check closer to where
it takes effect in order to reduce duplicated code. Also adds back in
integration tests to github actions.