Prevent KFP install timeouts during `stack up`
The Kubeflow Kustomize repository is quite large, as it consists of
several components. As a consequence, `kubectl apply -f` can sometimes
fail with a timeout while it's git-cloning the files locally. To prevent
that, a custom timeout value of 1 minute is used instead of the
kustomize default value of 27 seconds.

Signed-off-by: Stefan Nica <stefan@zenml.io>