Run KFP container as local user/group if local
Some files and folders in the local repositories (e.g. database,
.system TFX directories) are shared between different runs of the
same pipeline or of different pipelines. If pipelines are run
in the context of different users, such as when switching between
KFP and the local orchestrators, this can create problems with
permissions.

Running the KFP containers with the same UID/GID as the local
user maintains permissions consistent across orchestrators.

Signed-off-by: Stefan Nica <stefan@zenml.io>