Alexej/eng 300 bugfix windows integration tests (#296)
* Specified shell as executor and removed python3 specification

* Use shell=True in subprocess calls on Windows

This sets the value for the `shell=` parameter in subprocess calls to
click's belief whether or not the system is Windows. Setting it to True
alltogether completely breaks subprocess calls to paths with whitespace
on MacOS. This commit also moves the `if force` check closer to where
it takes effect in order to reduce duplicated code. Also adds back in
integration tests to github actions.

* Add integration tests to actions on PRs

* Make bash the default shell for github actions

* Specify shell executable via CLA/env variable

* adjust github env var syntax for bash

* Add shell executable variable to integration tests

* Make main and PR actions same

* Fix ml_flow integration test from main

* Add back bash shell and hotfix makedirs

* Prefix mlflow backend path with file:


Co-authored-by: Alexej Penner <thealexejpenner@gmail.com>