ZenML Profiles and Centralized ZenML Repository
* `zenml init` is optional. The global Repository is now initialized
automatically upon instantiation in the ~/.config/zenml location
* the Repository is made into a singleton, to avoid multiple instances
* the GlobalConfig is made into a singleton, to avoid multiple instances
* moved `version` from repository to global configuration, because it is
an attribute that needs to be addressed globally e.g. when migrating the
global configuration to a new version. Implemented rudimentary schema migration
and validation for the global configuration.
* implemented configuration profiles
* added support for building container images and loading source path classes
with dynamic detection of root python paths without a .zen repository root
Repository root is still optional, but takes precedence if present
* updated unit test and integration test fixtures to run without
initialized .zen repositories and use profiles instead
* move default stack creation to stack store initialization
* add `zenml profile explain` command
* fix all examples to make proper use of `main`
* add Profile to concepts page
* implemented automatic migration of legacy profiles
* parameterized the store back-end type in all unit and integration tests