Linking registered models in MLflow with the corresponding MLflow run (#3020)
* refactor: Add `mlflow_run_id` to the metadata used for `register_model_version`

The `mlflow_run_id` is required to assign a registered model to an MLflow run. The MLflow run can be traced back to a ZenML pipeline run.
This is not a necessary change. However, it is more convenient to have a reference to the MLflow run for the registered models as well, as this makes it possible to view parameters or metrics, for example.

* fix: Simplify and correct the if else statement

* fix: Remove unnecessary if