feat(integrations): Add MLX array materializer (#4027)
* feat(integrations): Add MLX array materializer

This enables saving and loading MLX arrays when using ZenML on Apple platforms.

Structurally, this is very similar to the recently added JAX integration, it's
just the array type that is different, plus we're using MLX's own array IO, so
there is no hard dependency on NumPy.

* tests: Add MLX array materializer test

Modeled after the existing tests, we test that the save-load round trip preserves
information and array values.

* typing: Force mypy to recognize the mx.load() output as single array

* fix: Restrict materializer to Apple + Linux, implement requirement hooks

Since Linux does not support Apple GPU computing, we need to install mlx with
the "cpu" extra on Linux platforms.